Initial Commit of BDCS API in Rust
Current state is that it is a Nickel.rs webapp with stubs for all of the
Composer demo v0 API endpoints. /api/v0/composer/types works.

It uses connection pooling so that the sqlite database can be shared
between API request threads.

CORS hasn't been enabled yet, but here's how:
https://github.com/nickel-org/nickel.rs/blob/master/examples/enable_cors.rs

The route matching is weird: pass foo.1 to /api/v0/dnf/info/ and it
matches but pass foo.1.1 and it doesn't match. Bug? Misunderstanding?