Add /api/v0/recipe/list
This is a *rough* draft:

It pulls recipe files from /var/tmp/recipes (because I don't know how to
pass configuration data down into the API function when it is called).

It needs better error handling.

BUT it does list the recipes, and /api/v0/recipe/<recipe file name>,...
will return the recipes in JSON format.

NOTE that the toml format has changed, see the example in ./examples/
The previous format wasn't valid.