Moving to Rocket.rs framework
This part writes the configuration settings to Rocket.toml
Ends up that Rocket *really* likes to use the config file for settings.
It has a ::custom() method, but if you use that you cannot access the
settings from the handlers.

Note that this requires using the nightly version of the rust compiler.

Basic framework is working, /api/v0/compose/types API is working.