Add /projects/list support
Returns a JSON object with information on the projects inside of
a {"projects":[]} list.

Moves some structs around, keeping them closer to the code using them
makes it easier to read.

Use a Filter struct to capture the limit and offset query parameters.
This is a bit quirky, rocket parses /projects/list as a different path
than /projects/list?offset=0&limit=20 so we need to set the rank on the
shorter one to a 2 and parse it with 2 different functions.