Add /recipes/list and /recipes/info/<name> API
Some changes to how it was being done before, behind the scenes the
recipe is stored as the 'name' but with spaces converted to '-' and
ending in .toml to make it clear to humans what is in the file.

Paths to the recipe API have changed to be more consistent with the
others. eg. /list/ to list things /info/ to get details about a specific
list of recipes.