Fix /recipes/ problems found by adding tests.
Result for /recipes/info/ should be a list, not a map.
Sort, dedupe and limit the results. Offset isn't implemented yet.

Move the RocketToml and RocketConfig into the bdcs library so that it
can be used in the tests.

Derive additional traits for the Recipe structs so that they can be
sorted and deduped.