Switch Dockerfile to use rustup nightly
rustup is the new tool for installing Rust, it makes it easy to switch
versions, etc. We need nightly for rocket, serde, etc.

Docker caches the result of the initial install so add a rustup update
to the final build command to actually update to the latest nightly
(some libraries won't build if the nightly is too far out of date).

So this may sometimes fail with:
info: update not yet available, sorry! try again later

That's not ideal, but I don't know what else to do until the features
needed by rocket and serde make it into stable.