Connect the recipes/new API to the git storage
Also change the recipe to use an empty array instead of Option for
modules and packages. Serde will handle setting these when parsing json
or toml thanks to the #[serde(default)] directive.

This commit requires using toml-rs version from the master branch until
there is a release (0.3.1 probably) that includes upstream commit
https://github.com/alexcrichton/toml-rs/commit/fba7778782193f0d0899f82d6b322da59a9044a0
which fixes a bug when converting toml to Value and then to a string.

Otherwise you may see a ValueAfterTable error at runtime, depending on
which array is empty.