Update tests
This is not an ideal situation. Rust has no way to provide global state
to the test threads, as well as no way to run a per-module
setup/teardown (or cleanup on success/failure). So this is clearly a
kludge until a better test framework can be written.

This stuffs all of the api tests into a single function, initializes the
database pool, sets up the recipe repository, and runs them all
single-threaded.

This still needs a copy of metadata.db to run, and
/var/tmp/bdcs-recipes-test/ is removed and recreated on every run.