Change depclose to return an Expression
Build an Expression from depclose instead of converting from Proposition to
Expression. This way the expression created by depclose is better able to
express the constraints caused by obsoletes/conflicts and by multiple providers
for a single requirement.

Version numbers are now checked during depclose, so the providers collected for
a given requirement all match the requirement.

This changes the type of Expression so that the atoms are either group ids or
parsed requirements. The pointer types within expressions are changed to
Rc<RefCell<DepExpression>> so that a group and its requirements can be used in
multiple sub-expressions without copying the data, and so the sub-expression
can later be modified during unit propagation.