Add recipe storage functions, API calls, and tests.
This adds the following API calls:

/recipes/changes/<recipes> - return the list of changes made to a recipe name
/recipes/diff/<recipe>/<from_commit>/<to_commit> - Show differences between two recipe changes.

It also adds new recipe functions for delete, commit log, and diff.

Tests for /recipes/changes/ /recipes/new/ with TOML recipe and
/recipe/diff have been added.

The initial recipe import has been changed to only import new recipes.
If one already exists in the repo it will be skipped to prevent
unnecessary commits, and overwriting any changes made to it.