Add another layer to the close_dependencies expression.
Each package requested in the packages Vec passed to close_dependencies
is something that must be installed, so the return should be an And() of
all of those packages. However, each request might have multiple groups
that satisfy it (multiarch, multiple versions, whatever), so each node
in the And is an Or of the possibilities for the particular package.