Add version number to the recipe
This uses semantic versioning rules (see http://www.semver.org) so that
the version can be incremented by the API.

If the new recipe is missing a version or has it set to "" it will
be set to 0.0.1

If the version in the new recipe is the same as the one in the last
commit, the patch number (z in x.y.z) will be incremented.

If the version in the new recipe doesn't match the last commit it will
be written as is (after checking to make sure it is a valid semver
version).

Errors in the new version will currently return a status of false.