If calling Requirement::from_str fails, use the whole string as the name.
The httpd package, in particular, provides "httpd-mmn =
20120211-x86-64".  According to from_str, this is not valid because the
release contains a dash.  If that happens, just take the entire input
string and use that as the name in the Requirement.  It is assumed that
even though such a string doesn't make sense to us, it'll make sense
when compared against requirements later.