Do not return multiple providers from close_dependencies.
Previously, any time something was provided by multiple packages we were
returning all possibilities and their dependency trees as one expression
OR'd together.  For some reason, this was causing unit_propagation to
get caught up in an endless loop trying to figure it all out.  This
problem was very difficult to debug because the expression is too large
to simply print out.

So instead, prevent multiple providers from being returned.  If there
are multiple options, simply do what we do in yum/dnf - choose the
package with the shortest name.