Fix the or_fun_call warnings from clippy.
The idea behind this warning is to make the code lazier: the argument to
unwrap_or() will be evaluated whether it's needed or not, so the
parameter passed to unwrap_or() is a function call there's code being
run unnecessarily.  All of our instances of function calls inside
unwrap_or parameters were to create an empty Vec or String, so we can
just use unwrap_or_default instead.