Add a manual Hash implementation for rpm::EVR.
Hash requires that, for any k1 and k2, k1 == k2 -> hash(k1) == hash(k2).
The problem with that as far as EVR is that we manually override Eq, so
that versions like "1|0|1" == "1.0.1", and if we then use the derived
Hash implementation those values would hash to different values.

To faciliate this, split the part of vercmp that divides a version into
components into a special Iterator implementation. The iterator provides
a sequence of normalized version components which can then be compared
one by one for vercmp or collected into a Vec and hashed for Hash.