Swith to Ubuntu Trusty on Travis CI, fixes SQLite errors
On RHEL 7 we have SQLite 3.7.11 which enhances the INSERT
syntax to allow multiple rows to be inserted via the VALUES clause:
https://www.sqlite.org/releaselog/3_7_11.html

The same syntax is used in `tests/db.rs` when creating the test
database. Older Travis containers use Ubuntu Precise, which has
SQLite 3.7.9 and thus produce an invalid syntax error!