Fix op-ref warnings
Clippy added a new lint. The idea behind this one is that an expression
like:

   (left_side: &T) == &(right_side: T)

is an unnecessary reference of the right-side value (to get it to match
the &T on the left). The compiler is going to have to dereference
everything eventually in order to compare the values, so it's better to
be upfront about it and write it as:

   *left_side == right_side