Support the new text in the key column of the KeyVal table.
The haskell-based importer serializes special data types that exist for
the key and writes that into the column.  Making that work in rust is a
pain and it's all temporary anyway, so don't do it that way.  Instead,
we just understand what haskell is writing into the database and
interpret it by treating it as strings and trimming off the data type
parts.

This also requires updating the tests and updating the test metadata.db
to include the new keys.

For now this just supports the TextKey type.