Include a bool in recipes/info output for workspace change
This adds a new list to the JSON response from /recipes/info:

"changes": [
    {
        "name": "recipe-test",
        "changed": true
    },
    {
        "name": "jboss",
        "changed": false
    }
],

If the recipe in the workspace is different than the most recent commit
the entry will have changed set to true. This list is next to the
recipes list in the JSON output:

{
    "changes": [ ..... ],
    "recipes": [ ..... ]
}

The recipes list has not changed.