Add finding recipe revision tags to the recipe::commits function
This will set the revision field to Some(x) if a tag, of the form
refs/tags/<branch>/<filename>/r<revition>, points to the commit.
Otherwise it is set to None.

Serde is configured to only return the revision field in the JSON
if it is not None so that it doesn't clutter up the results with
revision: null