Change the /recipes/diff/... response
This updates the recipe diff response to return the logical differences
between the recipes instead of line-by-line git style diffs.

It also adds the ability to compare the workspace recipe with a recipe
commit by using the WORKSPACE string in the to_commit hash field.

See the documentation for the recipes_diff function for details and an
example.