Install cargo build dependencies by hand
workaround around missing feature in cargo so that we can
create an intermediate container with build dependencies which
can be reused later to speed-up builds.

NOTE: some crates fail to build due to multiple versions available
(serde) or are a different version from what is required
(toml and chrono) but this is resolved during cargo build!