Run the tests inside Docker container
NOTE: moved COPY/ENTRYPOINT at the bottom of Dockerfile b/c
git checkout has newer timestamps and this is invalidating
intermediate Docker layers.

NOTE: upon successfull execution on the master branch the new
image is versioned with the TRAVIS_JOB_ID number and is pushed
to hub.docker.com. On the next iteration we pull this image and
try to use it as cache to speed up build jobs!