Update atlas recipe
atlas, lapack and blas are provide C/Fortran libraries. Either we
install the -devel pacakges in this recipe and try building some
math programs with them or we install numpy which depends on
atlas and lapack libraries and use it instead.

Btw numpy.test() is a nice side effect which we use in testing!