Dockerfile: drop the version number from clippy
This causes the build to fail when the compiler version is changed to
one that doesn't support the old version of clippy.

Installing without the version should be ok, it will be rebuilt when the
compiler nightly version is changed, and the clippy layer will remain
valid until the next compiler version.